Added separate component for Appbar
